feat: Adicionar funcionalidades de entrada de produto existente e atualizar lógica de cadastro de produtos

+++ database/migrations/2025_09_29_142942_create_secaos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('secaos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->unsignedInteger('fk_unidade');
+            $table->unsignedBigInteger('fk_unidade');
             $table->foreign('fk_unidade')->references('id')->on('unidades')->onDelete('cascade');
             $table->timestamps();
         });
